Move bookTrip and post functionality to domManipulation file
