+++ src/domManipulation.js
@@ -1,10 +1,21 @@
-// import { postData } from './fetch';
+import { postData } from './fetch';
+import { retrieveAllData } from './scripts';
 //
 const tripView = document.querySelector('#tripView')
 const todayDate = new Date().toISOString().slice(0, 10).replaceAll('-', '/')
 const destinationsSection = document.querySelector('#destinations')
 const bottomHeading = document.querySelector('#bottomHeading')
+const browsers = document.querySelector('#browsers')
+const browser = document.querySelector('#browser')
+const getEstimate = document.querySelector('#getEstimate')
+const estimateButton = document.querySelector('#estimateButton')
+const bookButton = document.querySelector('#bookButton')
+const glider = document.querySelector('#gliderSection')
+const startDate = document.querySelector('#startDate')
+const guests = document.querySelector('#guests')
+const numDays = document.querySelector('#numDays')
 
+////my account menu
 const changeFormView = (session) => {
   switch (tripView.value) {
   case 'past':
@@ -21,7 +32,6 @@ const changeFormView = (session) => {
     break;
   case 'expenses':
     let expenses = session.user.getTotalDollarSpentThisYear(session)
-    console.log('hello there')
     displayExpenses(expenses)
     break;
   default:
@@ -59,10 +69,70 @@ const displayExpenses = (expenses) => {
   '<h4>This year you have spent a total of $' + expenses + ' on fun excursions!</h4>'
 }
 
-//
+///////book trip section
+const addDestinationSearch = (session) => {
+  session.destinationData.forEach(dest => {
+    browsers.innerHTML += `
+    <option value="${dest.destination}">`
+  })
+}
+
+const showEstimate = (session) => {
+  if (browser.value && startDate.value && numDays.value && guests.value) {
+    let result = session.destinationObj.getTripCost(browser.value, numDays.value, guests.value)
+    getEstimate.innerHTML =
+    '<p class="estimate-calculation">Your estimate cost would be ' + '$' + result + '.</p>'
+    estimateButton.classList.add('hidden')
+    bookButton.classList.remove('hidden')
+  }
+}
+
+const bookTrip = (session) => {
+  const bookTripData = {
+    id: session.allTripsData.length + 1,
+    userID: session.user.id,
+    destinationID: getDestinationID(session),
+    travelers: guests.value,
+    date: formatDate(startDate.value),
+    duration: numDays.value,
+    status: 'pending',
+    suggestedActivities: []
+  }
+  postData(bookTripData)
+  retrieveAllData(session.user.id)
+}
+
+const getDestinationID = (session) => {
+  let result = session.destinationData.find(dest => {
+    if(dest.destination === browser.value) {
+      return dest
+    }
+  })
+  return result.id
+}
+
+const formatDate = (date) => {
+  const formattedDate = date.replaceAll('-', '/');
+  return formattedDate;
+}
+
+const addTripToPending = (location, date) => {
+  glider.classList.add('hidden')
+  destinationsSection.innerHTML += `
+  <section class='trips-display'>
+  <h4>${location}</h4>
+  <p>${date}</p>
+  </section>`
+}
+
 export {
   changeFormView,
   displayTrips,
-  displayExpenses
-
+  displayExpenses,
+  addDestinationSearch,
+  showEstimate,
+  bookTrip,
+  getDestinationID,
+  formatDate,
+  addTripToPending
 }
